erlang: Rename module interface file #82
