+++ src/modules/erlang/cnode.c
@@ -32,7 +32,7 @@
 
 #include "../../core/pass_fd.h"
 
-#include "mod_erlang.h"
+#include "erlang_mod.h"
 #include "erl_helpers.h"
 #include "cnode.h"
 #include "epmd.h"
